% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{conf_obj}
\alias{conf_obj}
\title{Confirmatory objective term with labels}
\usage{
conf_obj(y, z, D)
}
\arguments{
\item{y}{Treatment vector}

\item{z}{Lower dimension representation}

\item{D}{Original distance matrix}
}
\value{
list of objective values and sign (for computation later)
}
\description{
Confirmatory objective term with labels
}
