% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{get_dist_mat}
\alias{get_dist_mat}
\title{Distance between vectors}
\usage{
get_dist_mat(z)
}
\arguments{
\item{z}{Matrix or vector of observations}
}
\value{
Distance matrix
}
\description{
Distance between vectors
}
