% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{get_ind_mat}
\alias{get_ind_mat}
\title{Get index matrix}
\usage{
get_ind_mat(y)
}
\arguments{
\item{y}{a vector of treatments of length N}
}
\value{
A N by N matrix of indicators of equal treatment
}
\description{
Get index matrix
}
