% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{get_p}
\alias{get_p}
\title{Get p-value of PERMANOVA}
\usage{
get_p(mat = NULL, d = NULL, trt, n_iter = 999)
}
\arguments{
\item{mat}{Object matrix; used to build distance matrix d; d is prioritized}

\item{d}{Distance matrix; if NULL, obtain from mat using Euclidean distance}

\item{trt}{Vector of treatments}

\item{n_iter}{Number of iterations; defaults to 999}
}
\value{
list of ratio_all: vector of obtained pseudo-F values from permutations,
ratio: pseudo-F value, p: p-value from PERMANOVA
}
\description{
Get p-value of PERMANOVA
}
