% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmds.R
\name{getDistMat}
\alias{getDistMat}
\title{Distance between vectors}
\usage{
getDistMat(z)
}
\arguments{
\item{z}{Matrix or vector of observations}
}
\value{
Distance matrix
}
\description{
Distance between vectors
}
\examples{
set.seed(100)
z <- rbind(matrix(rnorm(100), ncol=4),
matrix(rnorm(100, 2), ncol=4))
getDistMat(z)
}
