% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{getIndMat}
\alias{getIndMat}
\title{Get index matrix}
\usage{
getIndMat(y)
}
\arguments{
\item{y}{a vector of treatments of length N}
}
\value{
A N by N matrix of indicators of equal treatment
}
\description{
Get index matrix
}
\examples{
require(phyloseq)
data(microbiome)
y <- sample_data(microbiome)$Treatment
getIndMat(y)
}
