% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{pair_by_rank}
\alias{pair_by_rank}
\title{Local regression on one-on-one paired by p value}
\usage{
pair_by_rank(D, z, y)
}
\arguments{
\item{D}{Original distance matrix}

\item{z}{Lower dimension representation}

\item{y}{Treatment vector}
}
\value{
matrix of pseudo-F values;
1st col of original data, 2nd col of reduced dim
}
\description{
Local regression on one-on-one paired by p value
}
