% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{pseudo_F}
\alias{pseudo_F}
\title{Title Get pseudo-F value for PERMANOVA}
\usage{
pseudo_F(mat = NULL, trt, d = NULL)
}
\arguments{
\item{mat}{Object matrix; used to build distance matrix d; d is prioritized}

\item{trt}{Vector of treatments}

\item{d}{Distance matrix; if NULL, obtain from mat using Euclidean distance}
}
\value{
pseudo-F value
}
\description{
Title Get pseudo-F value for PERMANOVA
}
