% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmds.R
\name{fmds}
\alias{fmds}
\title{FMDS calculation using MM algorithm}
\usage{
fmds(nit = 100, lambda = 0.2, threshold = 0.01, z0 = NULL, D, y, z)
}
\arguments{
\item{nit}{Number of iterations; 100 by default}

\item{lambda}{Hyperparameter; 0.2 by default}

\item{threshold}{Lower limit of p-value difference that allows iteration}

\item{z0}{Initialization of configuration; NULL by default}

\item{D}{Original distance matrix}

\item{y}{Treatment vector}

\item{z}{Object matrix; used to build distance matrix d; d is prioritized}
}
\value{
list of z: obtained configuration,
F_z: obtained pseuo-F value, F_0: original pseudo-F value
}
\description{
FMDS calculation using MM algorithm
}
\examples{
set.seed(100)
z0 <- cmdscale(d = microbiome$dist)
fmds(z0 = z0, D = microbiome$dist, y = microbiome$host)
}
