% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmds.R
\name{fmds}
\alias{fmds}
\title{FMDS calculation using MM algorithm}
\usage{
fmds(nit = 100, lambda = 0.5, threshold = 0.01, z0 = NULL, D, y, X)
}
\arguments{
\item{nit}{Number of iterations; 100 by default}

\item{lambda}{Hyperparameter; 0.5 by default}

\item{threshold}{Lower limit of p-value difference that allows iteration}

\item{z0}{Initialization of configuration; NULL by default}

\item{D}{Square matrix of pairwise distance, size of N by N}

\item{y}{Vector of label or group set, size of N}

\item{X}{Object matrix; used to build distance matrix D; D is prioritized}
}
\value{
2D representation vector, size of N by 2
}
\description{
FMDS calculation using MM algorithm
}
\examples{
set.seed(100)
z0 <- cmdscale(d = microbiome$dist)
fmds(z0 = z0, D = microbiome$dist, y = microbiome$host)
}
