% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mds_obj}
\alias{mds_obj}
\title{MDS objective}
\usage{
mds_obj(D, z)
}
\arguments{
\item{D}{Original distance matrix}

\item{z}{Lower dimension representation}
}
\value{
Scalar of objective function value of MDS
}
\description{
MDS objective
}
