% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmds.R
\name{mdsObj}
\alias{mdsObj}
\title{Objective term of MDS}
\usage{
mdsObj(D, z)
}
\arguments{
\item{D}{Original distance matrix}

\item{z}{Lower dimension representation}
}
\value{
Scalar of objective function value of MDS
}
\description{
Objective term of MDS
}
\examples{
require(phyloseq)
data(microbiome)
D <- distance(microbiome, method = 'wunifrac') # requires phyloseq package
y <- microbiome@sam_data@.Data[[1]]
z0 <- cmdscale(d = D)
mdsObj(D = D, z = z0)
}
