% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm_cmds}
\alias{mm_cmds}
\title{MM algorithm for FMDS}
\usage{
mm_cmds(nit = 100, lambda = 0.2, z0, D, y)
}
\arguments{
\item{nit}{Number of iterations; 100 by default}

\item{lambda}{Hyperparameter}

\item{z0}{Initialization of configuration; Can use MDS output or random init.}

\item{D}{Original distance matrix}

\item{y}{Treatment vector}
}
\value{
list of z: obtained configuration,
F_z: obtained pseuo-F value, F_0: original pseudo-F value
}
\description{
MM algorithm for FMDS
}
