% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm_cmds}
\alias{mm_cmds}
\title{MM algorithm for FMDS}
\usage{
mm_cmds(nit = 100, lambda = 0.2, z0, D, y)
}
\arguments{
\item{nit}{Number of iterations; 100 by default}

\item{lambda}{Hyperparameter}

\item{z0}{Initialization of lower dimension representation; Can use MDS result}

\item{D}{Original distance matrix}

\item{y}{Treatment vector}
}
\description{
MM algorithm for FMDS
}
